# Conflicts:
#	src/main/java/com/ttubeog/domain/area/domain/DongArea.java
#	src/main/java/com/ttubeog/domain/store/application/StoreService.java
#	src/main/java/com/ttubeog/domain/store/domain/Store.java
#	src/main/java/com/ttubeog/domain/store/dto/response/GetStoreDetailRes.java
#	src/main/java/com/ttubeog/domain/store/presentation/StoreController.java

+++ src/main/java/com/ttubeog/domain/store/application/StoreService.java
@@ -1,8 +1,16 @@
 package com.ttubeog.domain.store.application;
 
 import com.ttubeog.domain.area.domain.DongArea;
+import com.ttubeog.domain.area.domain.repository.DongAreaRepository;
+import com.ttubeog.domain.auth.security.JwtTokenProvider;
+import com.ttubeog.domain.image.application.ImageService;
+import com.ttubeog.domain.image.domain.Image;
+import com.ttubeog.domain.image.domain.ImageType;
+import com.ttubeog.domain.image.domain.repository.ImageRepository;
+import com.ttubeog.domain.image.dto.request.CreateImageRequestDto;
 import com.ttubeog.domain.member.domain.repository.MemberRepository;
 import com.ttubeog.domain.member.exception.InvalidMemberException;
+import com.ttubeog.domain.spot.exception.InvalidImageListSizeException;
 import com.ttubeog.domain.store.domain.Store;
 import com.ttubeog.domain.store.domain.repository.StoreRepository;
 import com.ttubeog.domain.store.dto.request.RegisterStoreReq;
@@ -10,16 +18,21 @@ import com.ttubeog.domain.store.dto.request.UpdateStoreReq;
 import com.ttubeog.domain.store.dto.response.GetStoreDetailRes;
 import com.ttubeog.domain.store.dto.response.RegisterStoreRes;
 import com.ttubeog.domain.store.dto.response.UpdateStoreRes;
+import com.ttubeog.domain.store.exception.InvalidDongAreaException;
 import com.ttubeog.domain.store.exception.UnathorizedMemberException;
 import com.ttubeog.domain.store.exception.NonExistentStoreException;
-import com.ttubeog.global.config.security.token.UserPrincipal;
 import com.ttubeog.domain.member.domain.Member;
 import com.ttubeog.global.payload.ApiResponse;
 import com.ttubeog.global.payload.Message;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+import static com.ttubeog.domain.image.application.ImageService.getImageString;
 
 @RequiredArgsConstructor
 @Service
@@ -28,25 +41,48 @@ public class StoreService {
 
     private final StoreRepository storeRepository;
     private final MemberRepository memberRepository;
+    private final DongAreaRepository dongAreaRepository;
+    private final ImageRepository imageRepository;
+    private final ImageService imageService;
+    private final JwtTokenProvider jwtTokenProvider;
 
     // 매장 등록
     @Transactional
-    public ResponseEntity<?> registerStore(UserPrincipal userPrincipal, RegisterStoreReq registerStoreReq) {
+    public ResponseEntity<?> registerStore(HttpServletRequest request, RegisterStoreReq registerStoreReq) {
 
-        Member member = memberRepository.findById(userPrincipal.getId()).orElseThrow(InvalidMemberException::new);
+        Long memberId = jwtTokenProvider.getMemberId(request);
+        Member member = memberRepository.findById(memberId).orElseThrow(InvalidMemberException::new);
+
+        DongArea dongArea = dongAreaRepository.findById(registerStoreReq.getDongAreaId()).orElseThrow(InvalidDongAreaException::new);
+
+        if (registerStoreReq.getImage().isEmpty() || registerStoreReq.getImage().size() > 10) {
+            throw new InvalidImageListSizeException();
+        }
 
         Store store = Store.builder()
                 .name(registerStoreReq.getName())
                 .info(registerStoreReq.getInfo())
-                .dongArea(new DongArea(registerStoreReq.getDongAreaId()))
+                .dongArea(dongArea)
+                .member(member)
                 .detailAddress(registerStoreReq.getDetailAddress())
                 .latitude(registerStoreReq.getLatitude())
                 .longitude(registerStoreReq.getLongitude())
-                .image(registerStoreReq.getImage())
+                .stars(0.0f)
                 .type(registerStoreReq.getType())
                 .build();
 
         storeRepository.save(store);
+
+        // 이미지 저장
+        List<String> imageList = registerStoreReq.getImage();
+        for (String s : imageList) {
+            CreateImageRequestDto createImageRequestDto = CreateImageRequestDto.builder()
+                    .image(s)
+                    .imageType(ImageType.SPOT)
+                    .placeId(store.getId())
+                    .build();
+            imageService.createImage(createImageRequestDto);
+        }
 
         RegisterStoreRes registerStoreRes = RegisterStoreRes.builder()
                 .storeId(store.getId())
@@ -57,7 +93,8 @@ public class StoreService {
                 .detailAddress(store.getDetailAddress())
                 .latitude(store.getLatitude())
                 .longitude(store.getLongitude())
-                .image(store.getImage())
+                .image(getImageString(imageRepository.findByStoreId(store.getId())))
+                .stars(store.getStars())
                 .type(store.getType())
                 .build();
 
@@ -71,14 +108,19 @@ public class StoreService {
 
     // 매장 수정
     @Transactional
-    public ResponseEntity<?> updateStore(UserPrincipal userPrincipal, UpdateStoreReq updateStoreReq) {
+    public ResponseEntity<?> updateStore(HttpServletRequest request, UpdateStoreReq updateStoreReq) {
 
-        Member member = memberRepository.findById(userPrincipal.getId()).orElseThrow(InvalidMemberException::new);
+        Long memberId = jwtTokenProvider.getMemberId(request);
+        memberRepository.findById(memberId).orElseThrow(InvalidMemberException::new);
         Store store = storeRepository.findById(updateStoreReq.getStoreId()).orElseThrow(NonExistentStoreException::new);
 
-        Member storeOwner = store.getMember();
-        if (storeOwner.getId() != member.getId()) {
+        Long storeOwnerId = store.getMember().getId();
+        if (!storeOwnerId.equals(memberId)) {
             throw new UnathorizedMemberException();
+        }
+
+        if (updateStoreReq.getImage().isEmpty() || updateStoreReq.getImage().size() > 10) {
+            throw new InvalidImageListSizeException();
         }
 
         store.updateName(updateStoreReq.getName());
@@ -86,8 +128,26 @@ public class StoreService {
         store.updateDetailAddress(updateStoreReq.getDetailAddress());
         store.updateLatitude(updateStoreReq.getLatitude());
         store.updateLongitude(updateStoreReq.getLongitude());
-        store.updateImage(updateStoreReq.getImage());
         store.updateType(updateStoreReq.getType());
+
+        storeRepository.save(store);
+
+        List<Image> imageList = imageRepository.findByStoreId(store.getId());
+
+        for (Image image : imageList) {
+            imageService.deleteImage(image.getId());
+        }
+
+        List<String> imageStringList = updateStoreReq.getImage();
+
+        for (String s : imageStringList) {
+            CreateImageRequestDto createImageRequestDto = CreateImageRequestDto.builder()
+                    .image(s)
+                    .imageType(ImageType.SPOT)
+                    .placeId(store.getId())
+                    .build();
+            imageService.createImage(createImageRequestDto);
+        }
 
         UpdateStoreRes updateStoreRes = UpdateStoreRes.builder()
                 .storeId(store.getId())
@@ -96,7 +156,8 @@ public class StoreService {
                 .detailAddress(store.getDetailAddress())
                 .latitude(store.getLatitude())
                 .longitude(store.getLongitude())
-                .image(store.getImage())
+                .image(getImageString(imageRepository.findByStoreId(store.getId())))
+                .stars(store.getStars())
                 .type(store.getType())
                 .build();
 
@@ -110,11 +171,23 @@ public class StoreService {
 
     // 매장 삭제
     @Transactional
-    public ResponseEntity<?> deleteStore(UserPrincipal userPrincipal, Long storeId) {
+    public ResponseEntity<?> deleteStore(HttpServletRequest request, Long storeId) {
 
-        memberRepository.findById(userPrincipal.getId()).orElseThrow(InvalidMemberException::new);
+        Long memberId = jwtTokenProvider.getMemberId(request);
+        memberRepository.findById(memberId).orElseThrow(InvalidMemberException::new);
         Store store = storeRepository.findById(storeId).orElseThrow(NonExistentStoreException::new);
+
+        Long storeOwnerId = store.getMember().getId();
+        if (!storeOwnerId.equals(memberId)) {
+            throw new UnathorizedMemberException();
+        }
+
         storeRepository.delete(store);
+
+        List<Image> imageList = imageRepository.findByStoreId(store.getId());
+        for (Image image : imageList) {
+            imageService.deleteImage(image.getId());
+        }
 
         ApiResponse apiResponse = ApiResponse.builder()
                 .check(true)
@@ -125,8 +198,10 @@ public class StoreService {
     }
 
     // 매장 세부사항 조회
-    public ResponseEntity<?> getStoreDetails(Long storeId) {
+    public ResponseEntity<?> getStoreDetails(HttpServletRequest request, Long storeId) {
 
+        Long memberId = jwtTokenProvider.getMemberId(request);
+        memberRepository.findById(memberId).orElseThrow(InvalidMemberException::new);
         Store store = storeRepository.findById(storeId).orElseThrow(NonExistentStoreException::new);
 
         // List<BenefitType> storeBenefits = benefitRepository.findTypeByStoreId(storeId);
@@ -142,7 +217,7 @@ public class StoreService {
                 .detailAddress(store.getDetailAddress())
                 .latitude(store.getLatitude())
                 .longitude(store.getLongitude())
-                .image(store.getImage())
+                .image(getImageString(imageRepository.findByStoreId(store.getId())))
                 .stars(store.getStars())
                 .type(store.getType())
                 //.storeBenefits(storeBenefits.stream().map(BenefitType::getType).collect(Collectors.toList()))
